requirements.txt
python-binance>=1.0.19
pyyaml>=6.0
python-dotenv>=1.0.0
pandas>=2.0.0
